@@ -2,7 +2,7 @@
   "name": "eslint-plugin-destructuring-newline",
   "version": "0.0.2",
   "description": "Enforce placing destructuring properties on separate lines",
-  "main": "lib/index.js",
+  "main": "dist/index.js",
   "repository": "git@github.com:kusotenpa/eslint-plugin-destructuring-newline.git",
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
@@ -16,13 +16,21 @@
     "newline"
   ],
   "scripts": {
-    "test": "mocha test/**/*.js --reporter spec"
+    "test": "jest",
+    "typecheck": "tsc",
+    "build": "swc ./src -d dist",
+    "upgrade-interactive": "npx npm-check-updates -i --format group"
   },
   "devDependencies": {
-    "@kusotenpa/eslint-plugin": "1.1.0",
-    "eslint": "7.6.0",
-    "eslint-plugin-rulesdir": "0.1.0",
-    "mocha": "8.1.1"
+    "@kusotenpa/eslint-plugin": "4.0.1",
+    "@swc/cli": "0.1.57",
+    "@swc/core": "1.3.2",
+    "@swc/jest": "0.2.22",
+    "@types/eslint": "8.4.6",
+    "eslint": "8.23.1",
+    "eslint-plugin-local-rules": "1.3.2",
+    "jest": "29.0.3",
+    "typescript": "4.8.3"
   },
   "peerDependencies": {
     "eslint": ">=7.0.0"
